test(server): split oversized registry tests
(cherry picked from commit 84e3e57202362c18334abbeba66d733e7bf288b4)
